Move api-client into app/
