+++ package.json
@@ -50,16 +50,8 @@
     "nodemon": "^1.11.0"
   },
   "nodemonConfig": {
-    "ignore":  [
-        ".git",
-        "test/*",
-        "node_modules/**/*",
-        "static/**/*"
-    ],
-    "watch": [
-        "app/",
-        "lib/"
-    ],
+    "ignore":  ["test/", "static/"],
+    "watch": ["app/"],
     "ext": "js json scss sass css html",
     "delay": "500",
     "verbose": true
